add ex 2

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -5,16 +5,17 @@
  * array_iterator - executes pointed function on each element of array
  * @array: array of elements
  * @size: size of array
+ * @action: action to perform on elements of array
  * Return: void
  */
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
-	
+	size_t i;
+
 	if (!array || !action)
 		return;
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
 		(*action)(*array);
 		array++;
